requirements.txt
haystack-ai
qdrant-haystack
sentence-transformers>=3.0.0 
huggingface_hub>=0.23.0
markdown-it-py 
mdit_plain 
pypdf
transformers[torch,sentencepiece]
python-dotenv
pymupdf
streamlit